+++ lib/core/null_displayer.cpp
@@ -63,3 +63,14 @@ void null_displayer::show_call_graph(const iface::call_graph&)
 {
   // throw away
 }
+
+void null_displayer::show_filename_list(
+    const std::vector<boost::filesystem::path>&)
+{
+  // throw away
+}
+
+void null_displayer::show_filename_set(const std::set<boost::filesystem::path>&)
+{
+  // throw away
+}