+++ lib/core/pragma_handler_map.cpp
@@ -37,6 +37,9 @@
 #include <metashell/pragma_echo.hpp>
 #include <metashell/pragma_macros.hpp>
 #include <metashell/pragma_macro_names.hpp>
+#include <metashell/pragma_includes.hpp>
+#include <metashell/pragma_which.hpp>
+#include <metashell/pragma_included_headers.hpp>
 
 #include <cassert>
 #include <sstream>
@@ -105,6 +108,21 @@ namespace
     return pragma_macro(
         "Set Metashell to " + name_ + " mode", move(cmds), shell_);
   }
+
+  template <class ForwardIterator, class Pred>
+  ForwardIterator
+  find_last_if(ForwardIterator begin_, ForwardIterator end_, Pred pred_)
+  {
+    ForwardIterator last = end_;
+    for (ForwardIterator i = begin_; i != end_; ++i)
+    {
+      if (pred_(*i))
+      {
+        last = i;
+      }
+    }
+    return last;
+  }
 }
 
 void pragma_handler_map::process(const data::command::iterator& begin_,
@@ -136,7 +154,14 @@ void pragma_handler_map::process(const data::command::iterator& begin_,
   if (longest_fit_handler)
   {
     longest_fit_handler->run(
-        begin_, longest_fit_begin, longest_fit_begin, e, displayer_);
+        begin_, find_last_if(begin_, longest_fit_begin,
+                             [](const data::token& token_)
+                             {
+                               return token_.category() !=
+                                      data::token_category::whitespace;
+                             }) +
+                    1,
+        longest_fit_begin, e, displayer_);
   }
   else
   {
@@ -232,6 +257,10 @@ pragma_handler_map::build_default(shell& shell_,
       .add("echo", pragma_echo())
       .add("macros", pragma_macros(shell_))
       .add("macro", "names", pragma_macro_names(shell_))
+      .add("sysincludes", pragma_includes<data::include_type::sys>(shell_))
+      .add("quoteincludes", pragma_includes<data::include_type::quote>(shell_))
+      .add("which", pragma_which(shell_))
+      .add("included", "headers", pragma_included_headers(shell_))
       .add("quit", pragma_quit(shell_));
 }